Move more Sarama YAML related functions into common
